refactor: client-node package with UnencryptedFileSystemKeyStore
