+++ packages/client-node/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@near-js/client-node",
+  "version": "0.0.1",
+  "description": "Dependencies for the NEAR API JavaScript client in NodeJS",
+  "main": "lib/index.js",
+  "scripts": {
+    "build": "pnpm compile",
+    "compile": "tsc -p tsconfig.json",
+    "test": "jest test"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@near-js/client-core": "workspace:*"
+  },
+  "devDependencies": {
+    "@types/node": "^18.7.14"
+  }
+}
