+++ packages/client-node/src/index.ts
@@ -1,0 +1,1 @@
+export * from './key_store';
