+++ packages/client-node/src/key_store/index.ts
@@ -1,0 +1,1 @@
+export { readKeyFile, UnencryptedFileSystemKeyStore } from './unencrypted_file_system_keystore';
